requirements.txt
asgiref==3.8.1
Django==5.2.1
django-spellbook==0.1.14b10
django-taggit==6.1.0
gunicorn==23.0.0
Markdown==3.8
packaging==25.0
python-decouple==3.8
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
